requirements.txt
djangorestframework==3.11.0
Django==3.1.2